# WARNING: head commit changed in the meantime
드라이버탈퇴-종학
메인헤더변경 - 회원가입 로그인 화면 안으로 넣음.
드라이버 mypage - 수정 중

+++ src/main/java/what/the/bus/driver/controller/DeleteDriverController.java
@@ -1,21 +1,32 @@
 package what.the.bus.driver.controller;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.SessionAttributes;
 
 import what.the.bus.driver.DriverVO;
 import what.the.bus.driver.service.DeleteDriverService;
 
 @Controller
+@SessionAttributes("member")
 public class DeleteDriverController {
 	@Autowired
 	private DeleteDriverService driverService;
 
 	// 회원탈퇴
 	@RequestMapping("/view/**/deleteDriver.do")
-	public String deleteDriver(DriverVO vo) {
-		driverService.deleteDriver(vo);
-		return "redirect:../main/main.jsp";
+	public String deleteDriver(@ModelAttribute("member") DriverVO vo, HttpSession session) {
+		if (driverService.pwCheckDriver(vo) == true) {
+			session.invalidate();
+			driverService.deleteDriver(vo);
+			return "redirect:../main/main.jsp";
+		} else {
+			return "redirect:../driver/errorPW.jsp";
+		}
+	
 	}
 }
